Move get_tangent_linear into a separate module, add documentation
