+++ tlm_adjoint/firedrake/firedrake_equations.py
@@ -28,7 +28,8 @@ from .backend_code_generator_interface import assemble, complex_mode, \
 
 from ..caches import Cache
 from ..equation import Equation
-from ..equations import ZeroAssignment, get_tangent_linear
+from ..equations import ZeroAssignment
+from ..tangent_linear import get_tangent_linear
 
 from .caches import form_dependencies, form_key, parameters_key
 from .equations import EquationSolver, bind_form, derivative, unbind_form, \
