+++ tlm_adjoint/overloaded_float.py
@@ -28,8 +28,9 @@ from .interface import DEFAULT_COMM, FunctionInterface, SpaceInterface, \
 
 from .caches import Caches
 from .equation import Equation
-from .equations import Assignment, ZeroAssignment, get_tangent_linear
+from .equations import Assignment, ZeroAssignment
 from .manager import annotation_enabled, tlm_enabled
+from .tangent_linear import get_tangent_linear
 
 from collections.abc import Sequence
 import contextlib
